Firing phasers now subtracts from firing Ship's energy. Tests to verify behavior. Resourced out more console text.

+++ trunk/StarTrek_KG/Interfaces/ISubsystem.cs
@@ -10,7 +10,7 @@ namespace StarTrek_KG.Interfaces
     {
         Map Map { get; set; }
         int Damage { get; set; }
-        int MaxTransfer { get; set; }
+        double MaxTransfer { get; set; }
         int Energy { get; set; }
         
         ISubsystem For(Ship ship);
